Added "infoWindowTemplate" and "infoWindowParameters" options to map markers.
By using "infoWindowTemplate" one can specify a url to a html template to be loaded for a marker's info window. This template can have its own angular controller so it's nicely separated.

The other option can be used in conjunction with "infoWindowTemplate" to send some parameters to the loaded template. The value of "infoWindowParameters" will be set as a "parameters" property on the parent scope of the template controller's scope, e.g. if setting "infoWindowParameters: { foo : 'bar' }" then from the template's controller it would be possible to do "$scope.parameters.foo" and get "bar".