issue #1114 This is the best that I can come up with. There is no native support, nor any support to watch an elements visibility. Yes this can be done with JQuery but we are not supporting JQuery, we are supporting JQlite only.
In the end this it needs to be made clear that if a div that contains a
map is hidden from the get go; then it should refresh or resize the map.

This is actually stated in the google sdk. If you use ng-if you do not
have to worry about this as the element is actually removed and then
linking is restarted. However if you use ng-show it is there , just
hidden and messes up the map.