fix(marker): markers now update on model changes
motivation: markers update relies on a comparison between a cached (cloned) version of the
model and the reference model. Prior to this commit, the cached model is built using a shallow
copy of the reference model (using `_.extends`). Thus, both cached and reference model properties
are always identical and model properties changes, e.g. coordinates, are not reflected on the map.
In this commit, the cached model is built using a *deep-copy* using `_.clone(model,true)`.

solves issue #1350