check maptype properties before refresh
Proposed solution to item that was piggy-backed onto issue #1542.

Idea is to check equality on specific properties of options object -
then trigger the map refresh.

So if your maptype (e.g. cartodb layer) relies on additional
properties/functions that are subject to change - but do not require a
map refresh, you won't get the flickering effect.