fix(uiGmapFitHelper): check object.hasOwnProperty() before attempting to extend the map bounds
If additional functionality has been added to Array.prototype, the for loop ends up trying to extend bounds with an invalid object (the extended prototype function). With this fix, the service will verify that the marker is indeed a valid property on the parent array before attempting to extend map bounds.